pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Александр Мышкин <sir.myshkin@gmail.com>"]
readme = "README.md"
packages = [{include = "hexlet_code"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
gunicorn = "^20.1.0"
flake8 = "^6.0.0"
pytest = "^7.2.2"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.5"
validators = "^0.20.0"
pydotenv = "^0.0.7"
requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
